requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1
structlog==24.4.0
langgraph>=0.2.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
trafilatura>=1.8.0
beautifulsoup4>=4.12.0
lxml_html_clean>=0.1.0
langchain-openai>=0.1.0
markdown>=3.10
